add style to list of problems
