--- frontend/src/pages/Profile/ChatList/index.js
@@ -1,3 +1,0 @@
-import ChatList from "./ChatList";
-
-export default ChatList;
