+++ frontend/src/pages/Profile/Problems/index.js
@@ -1,0 +1,3 @@
+import Problems from "./Problems";
+
+export default Problems;
