+++ frontend/src/pages/Profile/Profile.jsx
@@ -1,14 +1,14 @@
 import React from "react";
 import NavBar from "../../components/NavBar";
 import ProfileInfo from "./ProfileInfo";
-import ChatList from "./ChatList";
+import Problems from "./Problems";
 
 const Profile = () => {
   return (
     <div>
       <NavBar lightMode={true} />
       <ProfileInfo />
-      <ChatList />
+      <Problems />
     </div>
   );
 };
