+++ frontend/src/pages/Profile/ProfileInfo/ProfileInfo.css
@@ -1,4 +1,5 @@
 .jumbotron {
+  background-color: white;
   width: 100%;
   margin-top: 20px;
   display: flex;
